+++ packages/eyes-sdk-core/lib/positioning/PositionMemento.js
@@ -11,7 +11,7 @@ class PositionMemento {
    *  The keys are the style keys from which each of the transforms were taken
    * @param {Location} state.position - current location to be saved
    */
-  constructor({transforms, position}) {
+  constructor({transforms, position} = {}) {
     if (transforms) this._transforms = transforms
     if (position) this._position = new Location(position)
   }