+++ packages/eyes-webdriverio-5/src/SpecWrappedElement.js
@@ -3,7 +3,7 @@ const LegacySelector = require('./LegacySelector')
 
 /**
  * Supported selector type
- * @typedef {string|LegacySelector} SupportedSelector
+ * @typedef {string|Function|LegacySelector} SupportedSelector
  */
 
 /**
@@ -31,15 +31,14 @@ const ELEMENT_ID = 'element-6066-11e4-a52e-4f735466cecf'
 module.exports = {
   isCompatible(element) {
     if (!element) return false
-    return element.value
-      ? Boolean(element.value[ELEMENT_ID] || element.value[LEGACY_ELEMENT_ID])
-      : Boolean(element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID])
+    return Boolean(element.elementId || element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID])
   },
   isSelector(selector) {
-    return TypeUtils.isString(selector) || selector instanceof LegacySelector
-  },
-  extractId(element) {
-    return element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID]
+    return (
+      TypeUtils.isString(selector) ||
+      TypeUtils.isFunction(selector) ||
+      selector instanceof LegacySelector
+    )
   },
   toSupportedSelector(selector) {
     if (TypeUtils.has(selector, ['type', 'selector'])) {
@@ -63,20 +62,20 @@ module.exports = {
     }
     return {selector}
   },
+  extractId(element) {
+    return element.elementId || element[ELEMENT_ID] || element[LEGACY_ELEMENT_ID]
+  },
   extractElement(element) {
-    const unwrapped = element.value ? element.value : element
     return {
-      [ELEMENT_ID]: this.extractId(unwrapped),
-      [LEGACY_ELEMENT_ID]: this.extractId(unwrapped),
+      [ELEMENT_ID]: this.extractId(element),
+      [LEGACY_ELEMENT_ID]: this.extractId(element),
     }
   },
   extractSelector(element) {
     return element.selector
   },
-  isStaleElementReferenceResult(result, wrapper) {
+  isStaleElementReferenceResult(result) {
     if (!result) return false
-    return result instanceof Error
-      ? result.seleniumStack && result.seleniumStack.type === 'StaleElementReference'
-      : result.value && result.selector && result.selector === wrapper.selector
+    return result instanceof Error && result.name === 'stale element reference'
   },
 }