+++ packages/eyes-webdriverio-5/test/it/SpecWrappedElement.spec.js
@@ -9,10 +9,10 @@ describe('SpecWrappedDriver', async () => {
 
   before(async () => {
     await chromedriver.start([], true)
-    driver = remote({
-      desiredCapabilities: {
+    driver = await remote({
+      capabilities: {
         browserName: 'chrome',
-        chromeOptions: {
+        'goog:chromeOptions': {
           args: ['disable-infobars', 'headless'],
         },
       },
@@ -20,23 +20,22 @@ describe('SpecWrappedDriver', async () => {
       port: 9515,
       path: '/',
     })
-    await driver.init()
     await driver.url('https://applitools.github.io/demo/TestPages/FramesTestPage/')
   })
 
   after(async () => {
-    await driver.end()
+    await driver.deleteSession()
     chromedriver.stop()
   })
 
   it('isCompatible(element)', async () => {
-    const {value: element} = await driver.element('div')
+    const element = await driver.findElement('css selector', 'div')
     const result = specs.isCompatible(element)
     assert.deepStrictEqual(result, true)
   })
 
-  it('isCompatible(elementResponse)', async () => {
-    const element = await driver.element('div')
+  it('isCompatible(extendedElement)', async () => {
+    const element = await driver.$('div')
     const result = specs.isCompatible(element)
     assert.deepStrictEqual(result, true)
   })
@@ -51,6 +50,13 @@ describe('SpecWrappedDriver', async () => {
     assert.deepStrictEqual(result, true)
   })
 
+  it('isSelector(function)', async () => {
+    const result = specs.isSelector(function() {
+      return this.document.querySelector('div')
+    })
+    assert.deepStrictEqual(result, true)
+  })
+
   it('isSelector(by)', async () => {
     const result = specs.isSelector(By.css('div'))
     assert.deepStrictEqual(result, true)
@@ -59,20 +65,6 @@ describe('SpecWrappedDriver', async () => {
   it('isSelector(wrong)', async () => {
     const result = specs.isSelector({isSelector: true})
     assert.deepStrictEqual(result, false)
-  })
-
-  it('extractId(element)', async () => {
-    const elementId = 'elementId'
-    const element = {'element-6066-11e4-a52e-4f735466cecf': elementId}
-    const result = specs.extractId(element)
-    assert.deepStrictEqual(result, elementId)
-  })
-
-  it('extractId(legacyElement)', async () => {
-    const elementId = 'elementId'
-    const element = {ELEMENT: elementId}
-    const result = specs.extractId(element)
-    assert.deepStrictEqual(result, elementId)
   })
 
   it('toSupportedSelector(eyesSelector)', async () => {
@@ -107,8 +99,22 @@ describe('SpecWrappedDriver', async () => {
     assert.deepStrictEqual(wrongResult, {selector: wrongSelector})
   })
 
+  it('extractId(element)', async () => {
+    const elementId = 'elementId'
+    const element = {'element-6066-11e4-a52e-4f735466cecf': elementId}
+    const result = specs.extractId(element)
+    assert.deepStrictEqual(result, elementId)
+  })
+
+  it('extractId(legacyElement)', async () => {
+    const elementId = 'elementId'
+    const element = {ELEMENT: elementId}
+    const result = specs.extractId(element)
+    assert.deepStrictEqual(result, elementId)
+  })
+
   it('extractElement(element)', async () => {
-    const {value: element} = await driver.element('div')
+    const element = await driver.findElement('css selector', 'div')
     const elementId = element.ELEMENT || element['element-6066-11e4-a52e-4f735466cecf']
     const result = specs.extractElement(element)
     assert.deepStrictEqual(result, {
@@ -117,9 +123,9 @@ describe('SpecWrappedDriver', async () => {
     })
   })
 
-  it('extractElement(elementResponse)', async () => {
-    const element = await driver.element('div')
-    const elementId = element.value.ELEMENT || element.value['element-6066-11e4-a52e-4f735466cecf']
+  it('extractElement(extendedElement)', async () => {
+    const element = await driver.$('div')
+    const elementId = element.elementId
     const result = specs.extractElement(element)
     assert.deepStrictEqual(result, {
       ELEMENT: elementId,
@@ -129,20 +135,20 @@ describe('SpecWrappedDriver', async () => {
 
   it('extractSelector(element)', async () => {
     const selector = 'div'
-    const {value: element} = await driver.element(selector)
+    const element = await driver.findElement('css selector', selector)
     const result = specs.extractSelector(element)
     assert.deepStrictEqual(result, undefined)
   })
 
-  it('extractSelector(elementResponse)', async () => {
+  it('extractSelector(extendedElement)', async () => {
     const selector = 'div'
-    const element = await driver.element(selector)
+    const element = await driver.$(selector)
     const result = specs.extractSelector(element)
     assert.deepStrictEqual(result, selector)
   })
 
   it('isStaleElementReferenceResult(err)', async () => {
-    const {value: element} = await driver.element('div')
+    const element = await driver.$('div')
     await driver.refresh()
     try {
       await driver.execute('return null', element)