+++ packages/eyes-webdriverio-5/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## Unreleased
+- updated to @applitools/eyes-sdk-core@10.0.1
+- updated to @applitools/visual-grid-client@13.7.6
 
 - added retry for requests failing with a 503 response status
 - added more devices for device emulation